refactor: drop silent protocol defaults, split EditorCommand, document EDITOR
- Remove ProcessGateway.runInteractive default that silently fell back to
  run(), so the compiler enforces that every conformer chooses TTY-vs-piped
  semantics explicitly. UnimplementedProcessGateway now fatalErrors with the
  correct method name instead of misreporting "run not implemented".
- Remove ConfigHandler.editConfig/openConfig default failures for the same
  reason — UnimplementedConfigHandler already handles missing impls.
- Split the shell-style EditorCommand parser into its own file so
  ConfigHandlerImpl is just handler logic.
- README: document that $EDITOR is invoked as a literal "cmd args..." with
  no shell, ~, or $VAR expansion, and show the workaround for users who
  need shell features.
- Add explicit runInteractive stubs to the 6 existing ProcessGateway test
  doubles that were relying on the deleted default extension.
