+++ Tests/BenchmarkHandlerTests/BenchmarkHandlerTests.swift
@@ -35,6 +35,7 @@ private final class MockGateway: ProcessGateway, @unchecked Sendable {
     func runLaunchctl(_ arguments: [String]) -> Int32 { 0 }
     func findExecutable(_ name: String) -> String? { nil }
     func run(executable: String, arguments: [String]) -> Int32 { 0 }
+    func runInteractive(executable: String, arguments: [String]) -> Int32 { 0 }
     func runCapturingOutput(executable: String, arguments: [String]) -> String? { nil }
     func runStreaming(executable: String, arguments: [String]) -> AsyncStream<String> {
         AsyncStream { $0.finish() }
