+++ Tests/MediaRemoteDataSourceTests/MediaRemoteDataSourceImplTests.swift
@@ -188,6 +188,7 @@ private final class StreamingGateway: ProcessGateway, @unchecked Sendable {
     func runLaunchctl(_ arguments: [String]) -> Int32 { 0 }
     func findExecutable(_ name: String) -> String? { nil }
     func run(executable: String, arguments: [String]) -> Int32 { 0 }
+    func runInteractive(executable: String, arguments: [String]) -> Int32 { 0 }
     func runCapturingOutput(executable: String, arguments: [String]) -> String? { nil }
 
     func runStreaming(executable: String, arguments: [String]) -> AsyncStream<String> {
